Partial Objects to make full Transaction Object
