+++ src/transparent-qgw-database/transaction/options.ts
@@ -1,0 +1,43 @@
+import { DirectAPI, TransactionType } from "..";
+
+// Options:
+// emailCustomerReceipt: override_email_customer "Y" | "N";
+// sendTransactionEmail: override_trans_email "Y" | "N";
+// transctionType: trans_type
+// transactionID: transID
+// restrictKey: RestrictKey
+// dataSeparator: Dsep
+// maxMindOn: MAXMIND "1" | "2";
+export class Options {
+  public fields: OptionsFields;
+  constructor(
+    public emailCustomerReceipt: boolean = true,
+    public sendTransactionEmail: boolean = true,
+    public transctionType: TransactionType,
+    public transactionID: string,
+    public restrictKey?: string,
+    public dataSeparator?: string,
+    public maxMindOn?: boolean
+  ) {
+    this.fields = {
+      override_email_customer: emailCustomerReceipt ? "Y" : "N",
+      override_trans_email: sendTransactionEmail ? "Y" : "N",
+      trans_type: transctionType,
+      transID: transactionID,
+      RestrictKey: restrictKey,
+      Dsep: dataSeparator,
+      MAXMIND: maxMindOn ? "1" : "2",
+    };
+  }
+}
+
+export type OptionsFields = Pick<
+  DirectAPI,
+  | "override_email_customer"
+  | "override_trans_email"
+  | "trans_type"
+  | "transID"
+  | "RestrictKey"
+  | "Dsep"
+  | "MAXMIND"
+>;
